requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cachetools==4.2.2
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.1
h2==2.6.2
hpack==3.0.0
hyper==0.7.0
hyperframe==3.2.0
idna==3.3
lxml==4.9.1
multidict==6.0.2
python-dotenv==0.20.0
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.2
typing_extensions==4.4.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
yarl==1.8.1
